backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
supabase==2.4.6
openai==1.30.1
pypdf==4.2.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
slowapi==0.1.9

# Dev / test
pytest==8.2.2
pytest-asyncio==0.23.7

